Move things only used by the agent out of app.py and into the agent module
